requirements.txt
python-dotenv
Pillow
requests
psutil